+++ doc/source/changes.rst
@@ -8,6 +8,8 @@ Release 0.042 - ()
 * Add complex exponential integral.
 * Fix typo in cublasCgbmv.
 * Correctly load BLAS symbols on MacOSX and with CUDA 4.0.
+* Work around numpy bug #1898.
+* Fix issues with pycuda installations done via easy_install/pip. 
 
 Release 0.041 - (May 22, 2011)
 ------------------------------